搜索页内部布局，动态，话题，默认

+++ lib/data/model/data_response_model.dart
@@ -3,7 +3,7 @@ class  DataResponseModel {
   int hasNext;
   int lastTime;
   int lastId;
-  int lastScore;
+  double lastScore;
   DataResponseModel({this.list,this.hasNext, this.lastTime, this.lastId,this.lastScore});
 
   DataResponseModel.fromJson(Map<String, dynamic> json) {
